Fixed Ranged bug, finished tutorial level
Players ranged attack was aiming slightly above and to the left of the crosshair as that's where the mouse pointer would be using a regular cursor.

I altered some text hitboxes/timings and changed all slimes in the level to oozes. Hopefully no changes are needed to the tutorial level from this point onwards